Allow undefined BaseDirectoryType. Fix test.

+++ src/filesystem/TauriFileSystem.ts
@@ -15,6 +15,8 @@ import { exists, BaseDirectory, writeTextFile, readTextFile, writeBinaryFile, cr
  *  createDir(".", { dir: dir, recursive: true })   # does not work
  *  createDir("./", { dir: dir, recursive: true })  # does not work
  * Same for other API.
+ * 
+ * The file path selected by user via the tauri dialog api will be added to scope directly. 
  */
 export class TauriFileSystem implements IFileSystem {
     public static instance: TauriFileSystem = new TauriFileSystem()
@@ -31,41 +33,62 @@ export class TauriFileSystem implements IFileSystem {
         return path.substring(0, index)
     }
 
-    async exists(path: string, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<boolean> {
-        return exists(path, { dir: dir })
+    async exists(path: string, dir?: BaseDirectoryType): Promise<boolean> {
+        if (dir !== undefined){
+            return exists(path, { dir: dir })
+        }
+        return exists(path)
     }
 
-    async writeTextFile(path: string, content: string, append: boolean = true, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<void> {
+    async writeTextFile(path: string, content: string, append: boolean = false, dir?: BaseDirectoryType): Promise<void> {
         // If the path does not exists, writeTextFile will throw an error.
         let base_dir = this.baseDir(path)
         await this.createDirectory(base_dir, dir)
-        return writeTextFile(path, content, { dir: dir })
+        if (dir !== undefined){
+            return writeTextFile(path, content, { dir: dir, append: append })
+        }
+        return writeTextFile(path, content, { append: append })
     }
 
-    async readTextFile(path: string, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<string> {
-        return readTextFile(path, { dir: dir })
+    async readTextFile(path: string, dir?: BaseDirectoryType): Promise<string> {
+        if (dir !== undefined){
+            return readTextFile(path, { dir: dir })
+        }
+        return readTextFile(path)
     }
  
-    async writeBinaryFile(path: string, content: Uint8Array, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<void> {
+    async writeBinaryFile(path: string, content: Uint8Array, dir?: BaseDirectoryType): Promise<void> {
         // If the path does not exists, writeTextFile will throw an error.
         let base_dir = this.baseDir(path)
         await this.createDirectory(base_dir)
-        return writeBinaryFile(path, content, { dir: dir })
+        if (dir !== undefined){
+            return writeBinaryFile(path, content, { dir: dir })
+        }
+        return writeBinaryFile(path, content)
     }
 
-    async readBinaryFile(path: string, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<Uint8Array> {
-        return readBinaryFile(path, { dir: dir })
+    async readBinaryFile(path: string, dir?: BaseDirectoryType): Promise<Uint8Array> {
+        if (dir !== undefined){
+            return readBinaryFile(path, { dir: dir })
+        }
+        return readBinaryFile(path)
     }
 
-    async removeFile(path: string, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<void> {
-        return removeFile(path, { dir: dir })
+    async removeFile(path: string, dir?: BaseDirectoryType): Promise<void> {
+        if (dir !== undefined){
+            return removeFile(path, { dir: dir })
+        }
+        return removeFile(path)
     }
 
-    async createDirectory(path: string, dir: BaseDirectoryType = BaseDirectory.AppData): Promise<void> {
+    async createDirectory(path: string, dir?: BaseDirectoryType): Promise<void> {
         let exists = await this.exists(path, dir)
         if (!exists) {
             // create the directory if it does not exists
-            return createDir(path, { dir: dir, recursive: true })
+            if (dir !== undefined){
+                return createDir(path, { dir: dir, recursive: true })
+            }
+            return createDir(path, { recursive: true })
         }
         return Promise.resolve()
     }
